refactor controller
