+++ src/controllers/booking/index.ts
@@ -1,0 +1,4 @@
+export * from './create'
+export * from './list'
+export * from './update'
+export * from './delete'
