+++ src/controllers/organisation/index.ts
@@ -1,0 +1,4 @@
+export * from './catergory'
+export * from './list'
+export * from './delete'
+export * from './update'
