+++ src/controllers/user/index.ts
@@ -1,0 +1,4 @@
+export * from './create'
+export * from './delete'
+export * from './list'
+export * from './update'
