+++ src/routes/routes.ts
@@ -5,28 +5,8 @@ import {
   NextFunction,
   RequestHandler,
 } from 'express'
-import { handleLogin } from '@controllers/login'
 import { requiresAuthentication } from '@middlewares/auth.middleware'
-import {
-  getAllBookingsController,
-  getUserBookingsController,
-  createBooking,
-  editBooking,
-  deleteBookingHandler,
-} from '@controllers/bookings'
-import {
-  deleteOrganisation,
-  editOrganisation,
-  getOrgCategories,
-  getOrgs,
-} from '@controllers/organisations'
-import { getVenues } from '@controllers/venues'
-import {
-  createUser,
-  deleteUserFromDb,
-  editUser,
-  getUsers,
-} from '@controllers/users'
+import * as Controller from '@/controllers'
 
 export const router: Router = Router()
 
@@ -37,50 +17,70 @@ const asyncHandler =
   }
 
 // login route
-router.post('/login', asyncHandler(handleLogin))
+router.post('/login', asyncHandler(Controller.handleLogin))
 
 // get description of all Organisations
 // authentication not needed
-router.get('/orgs', asyncHandler(getOrgs))
-router.get('/orgs/categories', asyncHandler(getOrgCategories))
+router.get('/orgs', asyncHandler(Controller.listOrgs))
+router.get('/orgs/categories', asyncHandler(Controller.listOrgCategories))
 
 // create or edit an organisation
-router.put('/org/:id', requiresAuthentication, asyncHandler(editOrganisation))
+router.put(
+  '/org/:id',
+  requiresAuthentication,
+  asyncHandler(Controller.editOrganisation)
+)
 // delete an organisation
 router.delete(
   '/org/:id',
   requiresAuthentication,
-  asyncHandler(deleteOrganisation)
+  asyncHandler(Controller.deleteOrganisation)
 )
 
 // get all venues
-router.get('/venues', asyncHandler(getVenues))
+router.get('/venues', asyncHandler(Controller.listVenue))
 
 // get all users
-router.get('/users', requiresAuthentication, asyncHandler(getUsers))
+router.get('/users', requiresAuthentication, asyncHandler(Controller.listUser))
 
 // create a user
-router.post('/user', requiresAuthentication, asyncHandler(createUser))
+router.post(
+  '/user',
+  requiresAuthentication,
+  asyncHandler(Controller.createUser)
+)
 // edit a user
-router.put('/user/:id', requiresAuthentication, asyncHandler(editUser))
+router.put(
+  '/user/:id',
+  requiresAuthentication,
+  asyncHandler(Controller.editUser)
+)
 // delete a user
 router.delete(
   '/user/:id',
   requiresAuthentication,
-  asyncHandler(deleteUserFromDb)
+  asyncHandler(Controller.deleteUser)
 )
 
 // create a booking
-router.post('/bookings', requiresAuthentication, asyncHandler(createBooking))
+router.post(
+  '/bookings',
+  requiresAuthentication,
+  asyncHandler(Controller.createBooking)
+)
 // view all bookings
-router.get('/bookings/all', asyncHandler(getAllBookingsController))
+router.get('/bookings/all', asyncHandler(Controller.listBookings))
 // view bookings
-router.get('/bookings', asyncHandler(getUserBookingsController))
+router.get('/bookings', asyncHandler(Controller.getUserBookingsController))
 // edit a booking
-router.put('/bookings/:id', requiresAuthentication, asyncHandler(editBooking))
+router.put(
+  '/bookings/:id',
+  requiresAuthentication,
+  asyncHandler(Controller.editBooking)
+)
 // delete a booking
 router.delete(
   '/bookings/:id',
   requiresAuthentication,
-  asyncHandler(deleteBookingHandler)
+  asyncHandler(Controller.deleteBooking)
 )
